requirements.txt
# Qlib 量化投资平台学习项目依赖

# 核心依赖
pyqlib>=0.9.0
numpy>=1.19.0
pandas>=1.3.0
scipy>=1.7.0

# 机器学习模型
lightgbm>=3.3.0
scikit-learn>=1.0.0

# 数据处理
tables>=3.6.0

# 可选依赖（用于增强功能）
# pytorch>=1.8.0  # 如需使用深度学习模型
# redis>=3.5.0    # 用于缓存加速（可选）
